+++ apps/api/src/person/application/errors/index.ts
@@ -1,0 +1,2 @@
+export * from '@/person/application/errors/missing-field-error'
+export * from '@/person/application/errors/resource-not-found-error'
