+++ apps/api/src/person/domain/value-objects/index.ts
@@ -1,0 +1,2 @@
+export * from '@/person/domain/value-objects/document.value-object'
+export * from '@/person/domain/value-objects/email.value-object'
